Title: 최소공배수 2, Time: 12ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240425/최소공배수 2/least-common-multiple-2.cpp
@@ -15,7 +15,6 @@ int get_biggest_divisor(int a, int b) {
 
     for (int i = val; i > 0; i--) {
         if (a % i == 0 && b % i == 0) {
-            // cout << a << " % " << i << " = 0이고, " << b << " % " << i << " = 0이므로 최대공약수는 " << i << endl;
             return i;
         }
     }
@@ -27,8 +26,7 @@ int main() {
 
     int biggest_common_divisor = get_biggest_divisor(a, b);
     int divided_result = (a > b ? a : b) / biggest_common_divisor;
-    int result = (a > b ? b : a) * divided_result;
-
+    long result = long(a > b ? b : a) * long(divided_result);
     cout << result << endl;
     return 0;
 }
